refactor: Rename user atom file from atom to atoms
